Fix date/timezone bug where all ETAs are negative
If the current locale date is different to the current UTC date and the next event is the first one in the list, then this resulted in situations where all the event times showed as in the past (as they'd be using the previous day's date). So I've set it to use tomorrow's date if the ETA is negative, in order to force an ETA that is between 0 and one day.